Simulation AI: Implement looking ahead more than one move. This allows, for example, choosing the SpellAbility to use a fetchland to get the right type of mana in order to play another spell in hand, just from simulation.
This is an initial implementation with a max recursion depth of 5.

There are many improvements that can be made - for example to speed it up, by caching the planned SA chain - so that (if things are going to plan), the chain doesn't have to be recomputed each time.

Also, the AI doesn't currently simulated both types of decisions for a fetchland (i.e. evaluate difference between fetching a Plains vs. a Forest) - it picks one arbitrarily, which sometimes causes it to miss better plays. This needs to be fixed.