Fix card property controller checks to use LKI in case the card changed zones.
For example, this fixes the "It That Betrays" trigger that says "whenever an opponent sacrifices a nontoken permanent". Previously this would not work correctly - for example if the other player "It That Betrays" on the field and you sac an Evolving Wilds, they'll get control of that card (which is correct), but then when they sacrifice it, they would have gained control of it again (which is not correct).